Move round sig fig function to utilities
