+++ emutools/utils.py
@@ -9,6 +9,24 @@ from summer2.parameters import Function, Data, DerivedOutput
 from inputs.constants import INPUTS_PATH
 
 
+def round_sigfig(
+    value: float, 
+    sig_figs: int
+) -> float:
+    """
+    Round a number to a certain number of significant figures, 
+    rather than decimal places.
+    
+    Args:
+        value: Number to round
+        sig_figs: Number of significant figures to round to
+    """
+    if np.isinf(value):
+        return 'infinity'
+    else:
+        return round(value, -int(np.floor(np.log10(value))) + (sig_figs - 1)) if value != 0.0 else 0.0
+
+
 def triangle_wave_func(
     time: float, 
     start: float,
